pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "CrawJUD Backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "asn1crypto>=1.5.1",
    "bs4>=0.0.2",
    "celery>=5.5.3",
    "clear==2.0.0",
    "cryptography>=46.0.3",
    "dynaconf[all]>=3.2.12",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-keepass",
    "flask-mail>=0.10.0",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "httpx>=0.28.1",
    "hvac>=2.3.0",
    "hypercorn>=0.17.3",
    "inquirer>=3.4.1",
    "ip2location>=8.11.0",
    "jpype1>=1.6.0",
    "lxml>=6.0.2",
    "minio>=7.2.18",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib[argon2,bcrypt]>=1.7.4",
    "pg8000>=1.31.5",
    "pillow>=12.0.0",
    "py-bcrypt>=0.4",
    "pykeepass>=4.1.1.post1",
    "pyopenssl>=25.3.0",
    "pyotp>=2.9.0",
    "pypdf>=6.1.1",
    "pypsrp>=0.8.1",
    "python-dotenv==1.1.1",
    "python-socketio>=5.14.1",
    "pythonnet>=3.0.5",
    "quart>=0.20.0",
    "quart-socketio",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "rich==14.1.0",
    "selenium>=4.36.0",
    "selenium-wire>=5.1.0",
    "sqlalchemy==2.0.43",
    "termcolor==3.1.0",
    "tqdm==4.67.1",
    "typer>=0.20.0",
    "uvicorn>=0.37.0",
    "webdriver-manager>=4.0.2",
    "werkzeug>=3.1.5",
    "numpy>=2.4.0",
    "quart-cors",
    "websockets>=16.0",
    "quart-flask-patch",
    "anyio>=4.12.1",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.17",
    "pyarmor>=9.2.0",
    "pyarmor-man>=0.2",
    "pyinstaller>=6.16.0",
]

[tool.uv.sources]
flask-keepass = { git = "https://github.com/Robotz213/Flask-Keepass.git" }
quart-socketio = { git = "https://github.com/Robotz213/Quart-SocketIO.git" }
quart-cors = { git = "https://github.com/Robotz213/quart-flask-cors.git" }
quart-flask-patch = { git = "https://github.com/Robotz213/quart2-flask-patch.git" }
